Rename coconut editor mode
